videobot/requirements.txt
aiogram==3.7.0
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
